Added Rat in a Maze in C++ (Backtracking Algorithm)
The Rat in a Maze problem is a popular algorithmic problem in which a rat needs to find a path through a maze from a starting point to a destination. The maze is represented by a 2D grid, where each cell can either be open or blocked. The goal is to determine if there exists a path for the rat to reach the destination and, if so, we need to find such a path.

Approach:
Begin at the top-left corner of the maze.
From the current cell, explore moving in four possible directions (up, down, left, right).
Ensure that each move is valid (i.e., within maze boundaries and on an open cell).
Mark the current cell as part of the path if it is valid.
Recursively attempt to solve the maze from the next cell.
If the move does not lead to a solution, backtrack by unmarking the current cell and trying a different path.
If the destination is reached, the path is found.

Time Complexity: O(3^(m*n)), because on every cell we have to try 3 different directions , as we will not check for the cell from which we have visited in the last move.